pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bentwookie"
version = "0.1.0"
description = "AI coding loop workflow manager with SQLite state and Claude Agent SDK integration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "BentWookie Contributors"}
]
keywords = ["ai", "workflow", "task-management", "cli", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "questionary>=2.0.0",
    "flask>=3.0.0",
    "claude-agent-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
bw = "bentwookie.cli:main"

[project.urls]
Homepage = "https://github.com/bentwookie/bentwookie"
Repository = "https://github.com/bentwookie/bentwookie"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bentwookie = [
    "templates/**/*",
    "db/*.sql",
    "web/templates/*.html",
    "web/static/*.css",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
strict = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
